feat: 杜绝浪费，在Job上增加相关特性，避免不必要的开销 #970

+++ src/backend/ci/core/dispatch/api-dispatch/src/main/kotlin/com/tencent/devops/dispatch/pojo/JobQuotaStatus.kt
@@ -38,7 +38,7 @@ data class JobQuotaStatus(
     @ApiModelProperty("JOB告警阈值，百分比0-100", required = true)
     val jobThreshold: Int,
     @ApiModelProperty("项目job时间配额，单位：小时", required = true)
-    val timeQuota: Int,
+    val timeQuota: Long,
     @ApiModelProperty("当月已运行JOB时间, 单位：毫秒", required = true)
     val runningJobTime: Long,
     @ApiModelProperty("时间告警阈值，百分比0-100", required = true)
